Update docker-compose, revise example .env
Removed the hostname value from docker-compose as it occured to me
the simplest thing to do is to use localhost for redis and postgres
by default (from node service which runs locally). This compose file
is intended for local use anyways. example .env revised to allow
hostname to be set on install which will be updated to just be
127.0.0.1